requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0
google-generativeai>=0.8.0
dnspython>=2.6.0
python-dotenv>=1.0.0
